link-parser screen_width: Use casts
Prevent warnings when enabling -Wsign-conversion and -Wconversion
(or on Windows, when the equivalents are currently enabled).

+++ link-parser/command-line.c
@@ -853,7 +853,7 @@ static void put_opts_in_local_vars(Command_Options* copts)
 	local.use_sat_solver = parse_options_get_use_sat_parser(opts);
 	local.use_viterbi = parse_options_get_use_viterbi(opts);
 
-	local.screen_width = copts->screen_width;
+	local.screen_width = (int)copts->screen_width;
 	local.echo_on = copts->echo_on;
 	local.batch_mode = copts->batch_mode;
 	local.panic_mode = copts->panic_mode;
@@ -894,7 +894,7 @@ static void put_local_vars_in_opts(Command_Options* copts)
 	parse_options_set_use_viterbi(opts, local.use_viterbi);
 	parse_options_set_display_morphology(opts, local.display_morphology);
 
-	copts->screen_width = local.screen_width;
+	copts->screen_width = (size_t)local.screen_width;
 	copts->echo_on = local.echo_on;
 	copts->batch_mode = local.batch_mode;
 	copts->panic_mode = local.panic_mode;
